Convert tests from Tape to Jest
This change aligns with our module template.
