+++ package.json
@@ -10,7 +10,9 @@
   "main": "src/index.js",
   "types": "src/index.d.ts",
   "files": [
-    "src/"
+    "src/createEventEmitterProxy.js",
+    "src/createSwappableProxy.js",
+    "src/index.js"
   ],
   "scripts": {
     "lint": "yarn lint:eslint && yarn lint:misc --check && yarn lint:dependencies && yarn lint:changelog",
@@ -19,24 +21,28 @@
     "lint:eslint": "eslint . --cache --ext js",
     "lint:fix": "yarn lint:eslint --fix && yarn lint:misc --write && yarn lint:dependencies",
     "lint:misc": "prettier '**/*.json' '**/*.md' '!CHANGELOG.md' '**/*.yml' '!.yarnrc.yml' --ignore-path .gitignore --no-error-on-unmatched-pattern",
-    "test": "node test/"
+    "test": "jest && jest-it-up",
+    "test:watch": "jest --watch"
   },
   "devDependencies": {
     "@lavamoat/allow-scripts": "^2.0.3",
     "@metamask/auto-changelog": "^3.1.0",
     "@metamask/eslint-config": "^11.0.1",
+    "@metamask/eslint-config-jest": "^11.0.0",
     "@metamask/eslint-config-nodejs": "^11.0.1",
     "@typescript-eslint/parser": "^5.43.0",
     "depcheck": "^1.4.3",
     "eslint": "^8.27.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^27.1.5",
     "eslint-plugin-jsdoc": "^39.6.2",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^28.1.3",
+    "jest-it-up": "^2.0.2",
     "prettier": "^2.7.1",
-    "prettier-plugin-packagejson": "^2.3.0",
-    "tape": "^4.9.0"
+    "prettier-plugin-packagejson": "^2.3.0"
   },
   "packageManager": "yarn@3.4.1",
   "engines": {
