--- test/index.js
@@ -1,5 +1,0 @@
-// Requiring individual files is how Tape works.
-/* eslint-disable import/no-unassigned-import */
-
-require('./createSwappableProxy');
-require('./createEventEmitterProxy');
